+++ Code/diorama_private_files.cmake
@@ -7,9 +7,10 @@ set(FILES
     Source/Clients/SpriteComponent.cpp
     Source/Clients/SpriteComponent.h
     Source/Clients/SpriteComponentConfig.cpp
+    Source/Clients/SpriteAnimation.h
+    Source/Clients/SpriteBatchPlan.h
+    Source/Clients/SpriteFeatureProcessor.cpp
+    Source/Clients/SpriteFeatureProcessor.h
     Source/Clients/SpritePresenter.cpp
     Source/Clients/SpritePresenter.h
-    Source/Clients/SpriteRenderer.cpp
-    Source/Clients/SpriteRenderer.h
-    Source/Clients/SpriteRendererBus.h
 )